requirements.txt
Pillow==9.5.0
pydantic==1.10.9
fastapi==0.97.0
matplotlib==3.7.1
uvicorn==0.22.0
python-multipart==0.0.6
